+++ blackmamba/script/show_documentation.py
@@ -5,28 +5,18 @@ import editor
 import console
 import blackmamba.ide as ide
 import os
-import ui
 import jedi
 
 
-def _open_and_scroll(path, line):
-    if editor.get_path() == path:
-        ide.scroll_to_line(line)
-    else:
-        def scroll():
-            ide.scroll_to_line(line)
-        editor.open_file(path, new_tab=True)
-        ui.delay(scroll, 0.2)
-
-
 class LocationPickerItem(PickerItem):
-    def __init__(self, path, line, display_folder):
+    def __init__(self, path, line, display_folder, definition):
         super().__init__(
             '{}, line {}'.format(os.path.basename(path), line),
             display_folder
         )
         self.path = path
         self.line = line
+        self.definition = definition
 
     def matches_title(self, terms):
         if not terms:
@@ -61,9 +51,14 @@ class LocationDataSource(PickerDataSource):
         self.items = items
 
 
+def _show_documentation(definition):
+    line = ide.get_line_number()
+    editor.annotate_line(line, definition.docstring(), 'success')
+
+
 def _select_location(definitions):
-    def open_location(item, shift_enter):
-        _open_and_scroll(item.path, item.line)
+    def action(item, shift_enter):
+        _show_documentation(item.definition)
 
     v = load_picker_view()
     v.name = 'Multiple definitions found'
@@ -71,17 +66,17 @@ def _select_location(definitions):
 
     v.shift_enter_enabled = False
     v.help_label.text = (
-        '⇅ - select • Enter - open file and scroll to location'
+        '⇅ - select • Enter - show documentation'
         '\n'
         'Esc - close • Ctrl [ - close with Apple smart keyboard'
     )
     v.textfield.placeholder = 'Start typing to filter files...'
-    v.did_select_item_action = open_location
+    v.did_select_item_action = action
     v.present('sheet')
     v.wait_modal()
 
 
-def jump_to_definition():
+def show_documentation():
     path = editor.get_path()
     if not path:
         return
@@ -102,14 +97,14 @@ def jump_to_definition():
     definitions = script.goto_definitions()
 
     if not definitions:
-        console.hud_alert('Definition not found', 'error')
+        console.hud_alert('Documentation not found', 'error')
         return
 
     if len(definitions) == 1:
-        _open_and_scroll(definitions[0].module_path, definitions[0].line)
+        _show_documentation(definitions[0])
     else:
         _select_location(definitions)
 
 
 if __name__ == '__main__':
-    jump_to_definition()
+    show_documentation()